+++ genx/urls.py
@@ -4,7 +4,7 @@ from django.urls import path,include
 
 urlpatterns = [
     path('', include("core.urls")),
-    path('', include("courses.urls")),
+    path('courses/', include("courses.urls")),
     path('accounts/', include('allauth.urls')),
     path('admin/', admin.site.urls),
 ]